Moved incons to icons folder
